Added a cooldown to the shooting, made AI shoot and fixed a gameobject bug when adding a new one to the scene
The gun now has a cooldown and the AI can shoot bullets.  I had a bug where the game would crash if a bullet got spawned under influence by an update func. I fixed this by creating the lateadd in the scene. Which adds a gameobject after the update instead of during I also have to make a lateremove to solve the same crashes when deleting a go.